[TASK] Refactor DevLogDebugWriter to work with SolrLogManager
In #1086 the usage of devLog was replaced with the SolrLogManager that
uses the TYPO3 logging framework.

Since the setting tx_solr.plugin.logging.debugDevLogOutput does not make any sence
without the devLog it was renamed to tx_solr.plugin.logging.debugOutput.

To still output the log messages to the console (in the backend) or in the frontend when
tx_solr.plugin.logging.debugOutput is set to 1 the SolrLogManager needs to use the DebugWriter
to do that.

This PR:

* Renames the DevLogDebugWriter to DebugWriter
* Refactors the SolrLogManager to use the DebugWriter
* Adapts the existing tests

Impact:

When you set `tx_solr.plugin.logging.debugOutput = 1` the log messages in the frontend are
shown as debug message and in the backend in the TYPO3 console. When the setting is disabled
you do not see the log messages.

Fixes: #1090
